+++ client/packages/portal-client/src/utils/appsMatchingSearch.ts
@@ -1,0 +1,13 @@
+import { AppGroup } from '@equinor/portal-core';
+
+export function appsMatchingSearch(groups: AppGroup[], searchText?: string) {
+  if (!searchText) return groups;
+  return groups
+    .map((group) => ({
+      ...group,
+      applications: group.applications.filter((group) =>
+        group.appKey.toLowerCase().includes(searchText)
+      ),
+    }))
+    .filter((group) => group.applications.length);
+}
